+++ egui_demo_lib/src/apps/demo/mod.rs
@@ -19,7 +19,7 @@ pub mod scrolling;
 pub mod sliders;
 pub mod tests;
 pub mod toggle_switch;
-pub mod view_edit;
+pub mod password;
 pub mod widget_gallery;
 mod widgets;
 pub mod window_options;
